requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.6.0
django-daisy==0.1.28
djangorestframework==3.15.2
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.11.0
whitenoise==6.8.2
